🐛 fix: 재연결 시 인증 절차 재수행되도록 변경
