--- Backend/src/chats/chats.middleware.spec.ts
@@ -1,7 +1,0 @@
-import { ChatsMiddleware } from './chats.middleware';
-
-describe('ChatsMiddleware', () => {
-  it('should be defined', () => {
-    expect(new ChatsMiddleware()).toBeDefined();
-  });
-});
